+++ internal/metamorph/peer_handler.go
@@ -3,9 +3,9 @@ package metamorph
 import (
 	"context"
 	"fmt"
-	"github.com/bitcoin-sv/arc/pkg/tracing"
 
 	"github.com/bitcoin-sv/arc/internal/metamorph/store"
+	"github.com/bitcoin-sv/arc/internal/tracing"
 	"github.com/bitcoin-sv/arc/pkg/metamorph/metamorph_api"
 	"github.com/libsv/go-p2p"
 	"github.com/libsv/go-p2p/wire"
